refactor: improve timeline relations and controller logic
